updates in line with frameworkium-BDD changes
