+++ pom.xml
@@ -5,7 +5,7 @@
 
   <groupId>com.github.frameworkium</groupId>
   <artifactId>frameworkium-core</artifactId>
-  <version>2.4.1</version>
+  <version>2.4.2</version>
   <name>frameworkium-core</name>
   <description>
     Frameworkium core code.
@@ -66,6 +66,11 @@
       <groupId>ru.yandex.qatools.allure</groupId>
       <artifactId>allure-testng-adaptor</artifactId>
       <version>1.5.4</version>
+    </dependency>
+    <dependency>
+      <groupId>ru.yandex.qatools.allure</groupId>
+      <artifactId>allure-cucumber-jvm-adaptor</artifactId>
+      <version>1.6.3</version>
     </dependency>
     <dependency>
       <groupId>ru.yandex.qatools.htmlelements</groupId>
